ДОбавить получение задач для ADMIN, OWNER, добавить роли, добавить получение задач для конкретного пользователя, добавить завершение задачи.

+++ src/main/java/org/project/task/serviceImpl/TaskServiceImpl.java
@@ -8,11 +8,9 @@ import org.project.task.dto.response.task.TaskDto;
 import org.project.task.entity.Task;
 import org.project.task.mapper.task.MapperTask;
 import org.project.task.repository.TaskRepository;
-import org.project.task.service.GroupService;
 import org.project.task.service.GroupTasksService;
 import org.project.task.service.GroupUserService;
 import org.project.task.service.TaskService;
-import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.reactive.TransactionalOperator;
@@ -32,7 +30,6 @@ public class TaskServiceImpl implements TaskService {
 
     private final GroupUserService groupUserService;
 
-    private final GroupService groupService;
 
     private final GroupTasksService groupTasksService;
 
@@ -42,17 +39,27 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Mono<List<TaskDto>> getTasks(String timeZone, Jwt jwt, Long groupId) {
-        return groupService.verifyUserAccess(groupId, jwt).then(
-                taskRepository.findByGroupID(groupId).map(task -> mapperTask.taskToTaskDto(task, timeZone)).collectList());
+        return groupUserService.verifyAdminAccess(groupId, jwt).then(
+                taskRepository.findByGroupId(groupId).map(task -> mapperTask.taskToTaskDto(task, timeZone)).collectList());
     }
 
+    @Override
+    public Mono<List<TaskDto>> getTasksForUser(String timeZone, Jwt jwt, Long groupId) {
+        return groupUserService.validateUserInGroup(groupId, UUID.fromString(jwt.getSubject()))
+                .then(groupTasksService.findTaskForUserWithGroupId(jwt, groupId).collectList());
+    }
 
-
+    @Override
+    public Mono<Void> completeTask(Jwt jwt, Long taskId) {
+        return groupTasksService.findGroupIdByTaskID(taskId)
+                .flatMap(groupId -> groupTasksService.checkingWhetherUserIsPerformingThisTask(jwt, groupId, taskId)).then(
+                        taskRepository.updateFields(mapperTask.createCompleteTask(), Task.class, "id", taskId));
+    }
 
     @Override
     public Mono<Void> saveTask(Mono<CreateTaskDto> createTaskDtoMono, Jwt jwt) {
         return createTaskDtoMono
-                .flatMap(dto -> groupService.verifyUserAccess(dto.groupId(), jwt).thenReturn(dto))
+                .flatMap(dto ->  groupUserService.verifyAdminAccess(dto.groupId(), jwt).thenReturn(dto))
                 .flatMap(dto ->
                     taskRepository.save(mapperTask.taskDtoToTask(dto)).flatMap(task -> groupTasksService.saveTask(dto.groupId(), task.getId())
                     ).as(transactionalOperator::transactional).then()
@@ -62,7 +69,7 @@ public class TaskServiceImpl implements TaskService {
     @Override
     public Mono<Void> saveTaskByUserID(Mono<CreateTaskWithUserDto> createTaskWithUserDtoMono, Jwt jwt) {
         return createTaskWithUserDtoMono
-                .flatMap(dto ->  groupService.verifyUserAccess(dto.groupId(), jwt)
+                .flatMap(dto ->  groupUserService.verifyAdminAccess(dto.groupId(), jwt)
                         .thenReturn(dto))
                 .flatMap(dto -> groupUserService.validateUserInGroup(dto.groupId(), dto.userID()).thenReturn(dto))
                 .flatMap(dto ->
@@ -78,29 +85,18 @@ public class TaskServiceImpl implements TaskService {
     public Mono<Void> setTask(Mono<SetTaskDto> setTaskDtoMono, Jwt jwt) {
         return setTaskDtoMono
                 .flatMap(setTaskDto ->
-                    verifyUserAccess(setTaskDto.id(), jwt).thenReturn(mapperTask.createUpdateTaskFields(setTaskDto))
+                     groupTasksService.findGroupIdByTaskID(setTaskDto.id()).flatMap( groupId ->
+                    groupUserService.verifyAdminAccess(groupId, jwt).thenReturn(mapperTask.createUpdateTaskFields(setTaskDto))
                             .flatMap(sqlIdentifierObjectMap -> taskRepository.updateFields(sqlIdentifierObjectMap, Task.class, "id", setTaskDto.id()))
-                );
+                ));
     }
     @Override
     public Mono<Void> delTask(Long id, Jwt jwt) {
-        return verifyUserAccess(id, jwt)
-                .then(taskRepository.deleteById(id));
-    }
-
-
-    public Mono<Void> verifyUserAccess(Long taskId, Jwt jwt) {
-        return groupTasksService.existByTaskIdAndUserID(taskId, jwt).flatMap(bl -> {
-            if (!bl) {
-                return taskRepository.existsById(taskId).flatMap(ex -> {
-                            if (!ex) {
-                                return Mono.error(new NoSuchElementException("Элемент с id: " + taskId + " не найден"));
-                            }
-                            return Mono.error(new AccessDeniedException("У вас нет доступа работать с данной задачей"));
-                        }
+        return groupTasksService.findGroupIdByTaskID(id).flatMap(
+                groupID ->
+                groupUserService.verifyAdminAccess(groupID, jwt)
+                .then(taskRepository.deleteById(id))
                 );
-            }
-            return Mono.empty();
-        });
     }
+
 }
